+++ main.rb
@@ -1,29 +1,13 @@
 require_relative 'app'
+require_relative 'choice_method'
 
 def main
   app = App.new
   app.run
 end
 
-def options_list(option)
-  case option
-  when 1
-    list_book_display
-  when 2
-    list_people_display
-  when 3
-    create_person_display
-  when 4
-    create_book_display
-  when 5
-    create_rental_display
-  when 6
-    list_all_rental
-  else
-    puts 'Thanks for using the application'
+include ChoiceMethods
 
-  end
-end
 
 def display_menu
   puts 'Please choose an option by entering a number'
